Fix test of download option
After add seting up network step, the original test may be stucked at the the enter interface name step. So add an argument -network to indicate should or not set up network before download.
